Move Cardano related code to submodule
