+++ mithril-common/src/crypto_helper/mod.rs
@@ -1,9 +1,9 @@
 //! Tools and types to abstract the use of the [Mithril Core library](https://mithril.network/mithril-core/doc/mithril/index.html)
 
+mod cardano;
 mod codec;
 mod conversions;
 mod genesis;
-mod shelley_types;
 pub mod tests_setup;
 mod types;
